@@ -1,3 +1,3 @@
 module github.com/smartwalle/queue
 
-go 1.12
+go 1.18
