@@ -6,7 +6,7 @@ import (
 )
 
 func BenchmarkBlockQueue_Enqueue(b *testing.B) {
-	var q = block.New()
+	var q = block.New[int]()
 
 	for i := 0; i < b.N; i++ {
 		q.Enqueue(i)
