@@ -6,7 +6,7 @@ import (
 )
 
 func BenchmarkDelayQueue_Enqueue(b *testing.B) {
-	var q = delay.New()
+	var q = delay.New[int]()
 
 	for i := 0; i < b.N; i++ {
 		q.Enqueue(i, int64(i))
